fix watchlist and holdoing list duplicate rows bug

+++ src/main/java/com/wonder4/financeportfoliobackend/entity/UserWatchlist.java
@@ -23,7 +23,7 @@ public class UserWatchlist {
 
     private Long assetId;
 
-    @TableLogic private Integer isDeleted;
+    @TableLogic private Long isDeleted;
 
     @TableField(fill = FieldFill.INSERT)
     private LocalDateTime createTime;
